Remove hardcoded Kuberentes version.
By ommitting the kuberentesVersion field, in the kubeadm.conf, KubeAdm will
use the latest Kuberentes version that matches the KubeAdm version. For
example, if using KubeAdm for 1.10, it will use the latest (at that time
of invocation) of Kubernetes, say 1.10.4.

In README.md, suggesting that the user install CNI plugin 0.7.1+, which will
ensure that IPv6 support is present.

Note: The kubeadm.conf file in the work-area can be customized, after the
prepare step, if desired. For example, a specific Kubernetes version could
be specified.