Allow user configurable MTU for pod/management network
Sets the MTU used by CNI configuration file, based on user selection. Will
default to 1500, if not specified, and will require that MTU >= 1280, if
it is specified.

Also, separated out the validation logic for pod items and for DNS64 items.
Moved the setting of deprecated DNS flag into the validator, instead of
using in routine that calculates derived fields.