Version based kubeadm.conf files for 1.10, 1.11, 1.12, 1.13+. User can specify the exact Kubernetes version to be used, by setting the kubernetes-version field in the general section of the config.yaml file.
The value must begin with "v#.#.#" or be "latest", which will use the
latest Kuberentes code.  The version must have the same major and
minor version as the installed kubeadm tool.

Issues:

With 1.13 and using latest Kuberentes version, I see an error
at kubeadm init indicating that port 10250 is in use. Cheecking, I
see kubelet is using it. Difference between this and 1.12, is that
with 1.12, kubelet is constantly resetting, due to missing config
files and no port is in use. When kubeadm init runs, it creates the
files and then kubelet stays up. With 1.13, kubelet is already running
and using port 10250.

With 1.10, join fails saying that it cannot determine if container
runtime is running. This was brought up in the following issue, with
the disposition that it was fixed in 1.11 and not backported:

https://github.com/kubernetes/kubeadm/issues/814