Update for Kubernetes 1.13 and CoreDNS.
Since CoreDNS is the default, and the featureGate option is removed (as it
the --feature-gate CLI arg), the kubeadm template was updated to use the
new config syntax. In addition, there is a new configurable field, "TypeDNS"
that can be used for 1.13+. It is hard coded to CoreDNS for now, with no
YAML option to select yet (though the kubeadm.conf file could be manually
edited, after the "prepare" step).